# apps/api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "User Management API",
    "version": "1.0.0",
    "description": "API for user management and authentication"
  },
  "servers": [
    {
      "url": "{API}/api/{VERSION}",
      "variables": {
        "API": {
          "default": "http://localhost:3000"
        },
        "VERSION": {
          "default": "v1"
        }
      }
    }
  ],
  "paths": {
    "/users/email/{email}": {
      "get": {
        "summary": "Get user by email",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string"
                    },
                    "fullName": {
                      "type": "string"
                    },
                    "dob": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/users/countryCode/{countryCode}/phone/{phone}": {
      "get": {
        "summary": "Get user by phone",
        "parameters": [
          {
            "name": "countryCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phone",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "countryCode": {
                      "type": "string"
                    },
                    "phone": {
                      "type": "string"
                    },
                    "fullName": {
                      "type": "string"
                    },
                    "dob": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/users/is-there/email/{email}": {
      "get": {
        "summary": "Check if user email exists",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email existence status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "exist": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/is-there/countryCode/{countryCode}/phone/{phone}": {
      "get": {
        "summary": "Check if user phone exists",
        "parameters": [
          {
            "name": "countryCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phone",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Phone existence status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "exist": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/email/": {
      "post": {
        "summary": "Register user by email",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "fullName": {
                    "type": "string"
                  },
                  "dob": {
                    "type": "string"
                  }
                },
                "required": ["email", "fullName", "dob"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string"
                    },
                    "fullName": {
                      "type": "string"
                    },
                    "dob": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/users/phone/": {
      "post": {
        "summary": "Register user by phone",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "countryCode": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "fullName": {
                    "type": "string"
                  },
                  "dob": {
                    "type": "string"
                  }
                },
                "required": ["countryCode", "phone", "fullName", "dob"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "countryCode": {
                      "type": "string"
                    },
                    "phone": {
                      "type": "string"
                    },
                    "fullName": {
                      "type": "string"
                    },
                    "dob": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/auth/send-pin/email/": {
      "post": {
        "summary": "Send verification PIN to email",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  }
                },
                "required": ["email"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PIN sent successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/auth/send-pin/phone/": {
      "post": {
        "summary": "Send verification PIN to phone",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "countryCode": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  }
                },
                "required": ["countryCode", "phone"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PIN sent successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/auth/verification-pin/email/": {
      "post": {
        "summary": "Verify PIN for email",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "pin": {
                    "type": "string"
                  }
                },
                "required": ["email", "pin"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PIN verified successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/auth/verification-pin/phone/": {
      "post": {
        "summary": "Verify PIN for phone",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "countryCode": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "pin": {
                    "type": "string"
                  }
                },
                "required": ["countryCode", "phone", "pin"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PIN verified successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    }
  }
}